requirement.txt
pygame
pyaudio
pandas
pylsl
opencv-python
ffmpeg-python
numpy
pyglet
PyQt5
pyempatica
tobii_research